+++ example/lib/bloc/counter_dialog/bloc.dart
@@ -1,25 +1,30 @@
 import 'package:pure_flutter_bloc/bloc.dart';
-import 'state.dart';
-import 'events.dart';
+
 import 'change.dart';
+import 'events.dart';
+import 'state.dart';
 
 class CounterDialogBloc extends Bloc<CounterDialogState, CounterDialogEvent> {
   @override
   CounterDialogState initState() {
-    return CounterDialogState(
-      counterVal: 0
-    );
+    return CounterDialogState(counterVal: 0);
   }
 
   @override
   Stream<Change<CounterDialogState>> mapEventToChange(
       CounterDialogEvent event) async* {
     if (event is IncrementDialogCounterEvent) {
-      yield DialogCounterValChange(counterVal: 10, isIncrement: true);
+      yield DialogCounterValChange(
+        counterVal: 10,
+        isIncrement: true,
+      );
     }
 
     if (event is DecrementDialogCounterEvent) {
-      yield DialogCounterValChange(counterVal: 10, isIncrement: false);
+      yield DialogCounterValChange(
+        counterVal: 10,
+        isIncrement: false,
+      );
     }
   }
 }