+++ example/lib/bloc/counter_dialog/change.dart
@@ -1,21 +1,36 @@
+import 'dart:core';
+
 import 'package:flutter/foundation.dart';
 import 'package:pure_flutter_bloc/bloc.dart';
+
 import 'state.dart';
 
 class DialogCounterValChange extends Change<CounterDialogState> {
   final int counterVal;
   final bool isIncrement;
 
-  DialogCounterValChange({@required this.counterVal, @required this.isIncrement,});
+  Map<String, dynamic> toMap() {
+    // ignore: unnecessary_cast
+    return {
+      'counterVal': this.counterVal,
+      'isIncrement': this.isIncrement,
+    } as Map<String, dynamic>;
+  }
+
+  DialogCounterValChange({
+    @required this.counterVal,
+    @required this.isIncrement,
+  });
 
   @override
   CounterDialogState apply(CounterDialogState oldState) {
-    return oldState.copyWith(counterVal: isIncrement ? (oldState.counterVal + counterVal) : (oldState.counterVal - counterVal));
+    return oldState.copyWith(
+      counterVal: isIncrement
+          ? (oldState.counterVal + counterVal)
+          : (oldState.counterVal - counterVal),
+    );
   }
 
   @override
-  List<Map<String, Object>> get propsMap => [
-    {'counterVal': counterVal},
-    {'isIncrement': isIncrement},
-  ];
+  Map<String, Object> get propsMap => toMap();
 }