+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## [1.0.1] - 17/03/2021.
+
+* Breaking changes in subclasses that extends BlocEquatable class
+* Easy support to stringify BLoc **State** and **Change** instances.   
+
 ## [1.0.0] - 15/03/2021.
 
 * Initial version of the library.
 * Includes full support for implementing BLoc pattern using BlocWidget, Select, BlocProvider, MultiBlocProvider, BlocListener and MultiBlocListener
-  widgets.  
+  widgets.
+  
+